+++ Modular Gameplay Overhaul/Modules/Enchantments/Integrations/SpaceCoreIntegration.cs
@@ -25,13 +25,14 @@ internal sealed class SpaceCoreIntegration : ModIntegration<SpaceCoreIntegration
         this.AssertLoaded();
 
         // melee
-        this.ModApi.RegisterSerializerType(typeof(BloodthirstyEnchantment));
+        this.ModApi.RegisterSerializerType(typeof(NewArtfulEnchantment));
         this.ModApi.RegisterSerializerType(typeof(CarvingEnchantment));
         this.ModApi.RegisterSerializerType(typeof(CleavingEnchantment));
         this.ModApi.RegisterSerializerType(typeof(EnergizedEnchantment));
         this.ModApi.RegisterSerializerType(typeof(ExplosiveEnchantment));
-        this.ModApi.RegisterSerializerType(typeof(NewArtfulEnchantment));
         this.ModApi.RegisterSerializerType(typeof(TributeEnchantment));
+        this.ModApi.RegisterSerializerType(typeof(BloodthirstyEnchantment));
+        this.ModApi.RegisterSerializerType(typeof(WabbajackEnchantment));
 
         // ranged
         this.ModApi.RegisterSerializerType(typeof(BaseSlingshotEnchantment));