+++ Modular Gameplay Overhaul/Modules/Professions/Patchers/Common/CharacterInitNetFieldsPatcher.cs
@@ -25,8 +25,7 @@ internal sealed class CharacterInitNetFieldsPatcher : HarmonyPatcher
     {
         if (__instance is Farmer { Name: { } } farmer)
         {
-            __instance.NetFields.AddFields(farmer.Get_UltimateIndex());
-            __instance.NetFields.AddFields(farmer.Get_IsHuntingTreasure());
+            __instance.NetFields.AddFields(farmer.Get_UltimateIndex(), farmer.Get_IsHuntingTreasure());
         }
     }