+++ Modular Gameplay Overhaul/Modules/Slingshots/Events/SlingshotButtonPressedEvent.cs
@@ -25,7 +25,7 @@ internal sealed class SlingshotButtonPressedEvent : ButtonPressedEvent
     /// <inheritdoc />
     protected override void OnButtonPressedImpl(object? sender, ButtonPressedEventArgs e)
     {
-        if (!Context.IsWorldReady || Game1.activeClickableMenu is not null)
+        if (!Context.IsWorldReady || Game1.activeClickableMenu is not null || Game1.isFestival())
         {
             return;
         }