+++ Modular Gameplay Overhaul/Modules/Weapons/Patchers/MeleeWeaponDrawDuringUsePatcher.cs
@@ -73,8 +73,7 @@ internal sealed class MeleeWeaponDrawDuringUsePatcher : HarmonyPatcher
             var numFramesBeforeFinalHit = ((int)finalHitStep - 1) * 6;
             if (type == MeleeWeapon.club && (frameOfFarmerAnimation >= numFramesBeforeFinalHit || hitstep == finalHitStep))
             {
-                DrawDuringVerticalSwipe(
-                    frameOfFarmerAnimation,
+                DrawDuringSmash(
                     facingDirection,
                     spriteBatch,
                     playerPosition,
@@ -83,7 +82,7 @@ internal sealed class MeleeWeaponDrawDuringUsePatcher : HarmonyPatcher
             }
             else
             {
-                DrawDuringHorizontalSwipe(
+                DrawDuringSwipe(
                     ___center,
                     facingDirection,
                     spriteBatch,
@@ -241,7 +240,7 @@ internal sealed class MeleeWeaponDrawDuringUsePatcher : HarmonyPatcher
         }
     }
 
-    private static void DrawDuringHorizontalSwipe(
+    private static void DrawDuringSwipe(
         Vector2 center,
         int facingDirection,
         SpriteBatch spriteBatch,
@@ -249,11 +248,11 @@ internal sealed class MeleeWeaponDrawDuringUsePatcher : HarmonyPatcher
         Farmer farmer,
         Rectangle sourceRectangle)
     {
-        var frame = farmer.FarmerSprite.CurrentFrame;
+        var frameOfFarmerAnimation = farmer.FarmerSprite.CurrentFrame;
         switch (facingDirection)
         {
             case Game1.up:
-                switch (frame)
+                switch (frameOfFarmerAnimation)
                 {
                     case 36:
                         spriteBatch.Draw(
@@ -332,7 +331,7 @@ internal sealed class MeleeWeaponDrawDuringUsePatcher : HarmonyPatcher
                 break;
 
             case Game1.right:
-                switch (frame)
+                switch (frameOfFarmerAnimation)
                 {
                     case 30:
                         spriteBatch.Draw(
@@ -411,7 +410,7 @@ internal sealed class MeleeWeaponDrawDuringUsePatcher : HarmonyPatcher
                 break;
 
             case Game1.down:
-                switch (frame)
+                switch (frameOfFarmerAnimation)
                 {
                     case 24:
                         spriteBatch.Draw(
@@ -490,7 +489,7 @@ internal sealed class MeleeWeaponDrawDuringUsePatcher : HarmonyPatcher
                 break;
 
             case Game1.left:
-                switch (frame)
+                switch (frameOfFarmerAnimation)
                 {
                     case 30:
                         spriteBatch.Draw(
@@ -570,19 +569,18 @@ internal sealed class MeleeWeaponDrawDuringUsePatcher : HarmonyPatcher
         }
     }
 
-    private static void DrawDuringVerticalSwipe(
-       int frameOfFarmerAnimation,
-       int facingDirection,
-       SpriteBatch b,
-       Vector2 playerPosition,
-       Farmer farmer,
-       Rectangle sourceRect)
+    private static void DrawDuringSmash(
+        int facingDirection,
+        SpriteBatch b,
+        Vector2 playerPosition,
+        Farmer farmer,
+        Rectangle sourceRect)
     {
-        var frame = farmer.FarmerSprite.CurrentFrame;
+        var frameOfFarmerAnimation = farmer.FarmerSprite.CurrentFrame;
         switch (facingDirection)
         {
             case Game1.right:
-                switch (frame)
+                switch (frameOfFarmerAnimation)
                 {
                     case 35:
                         b.Draw(
@@ -662,7 +660,7 @@ internal sealed class MeleeWeaponDrawDuringUsePatcher : HarmonyPatcher
                 break;
 
             case Game1.left:
-                switch (frame)
+                switch (frameOfFarmerAnimation)
                 {
                     case 35:
                         b.Draw(
@@ -742,7 +740,7 @@ internal sealed class MeleeWeaponDrawDuringUsePatcher : HarmonyPatcher
                 break;
 
             default:
-                switch (frame)
+                switch (frameOfFarmerAnimation)
                 {
                     // remnants of first hit facing down
                     case 29: